multi_rag/requirements_streamlit.txt
streamlit
python-dotenv
langchain
langchain-community
langchain-core
langchain-openai
langchain-anthropic
langchain-groq
faiss-cpu
PyPDF2
tiktoken
